auth-service/sample-api-users/requirements.txt
## Prod
flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
flask-cors==3.0.10
flask-bcrypt==1.0.1
pyjwt==2.6.0
gunicorn==20.1.0